requirements.txt
requests~=2.32.2
python-dotenv~=1.0.1
pytelegrambotapi~=4.18.1